refactor: move ValueMemberAccess enum to FieldDescriptor
Moved ValueMemberAccess enum from ValueMemberFieldDescriptor to FieldDescriptor
to make it reusable across all field descriptor types.

Changes:
- Moved ValueMemberAccess enum to FieldDescriptor class
- Made ValueMemberAccess.value field public for direct access
- Replaced boolean isPublic with ValueMemberAccess valueMemberAccess
- Updated FieldDescriptor constructor to set valueMemberAccess
- Updated getValueMember() to use valueMemberAccess.value
- Added access field to ValueMemberFieldDescriptor
- Fixed RemoteFieldDescriptor to properly initialize final interfaceType
- Added static import for ValueMemberAccess in ValueMemberFieldDescriptor

Benefits:
- Better type safety with enum vs boolean
- Clearer intent with named access levels
- Reusable across all FieldDescriptor subclasses
- Proper initialization of final fields
