+++ yoko-rmi-impl/src/main/java/org/apache/yoko/rmi/impl/ValueMemberFieldDescriptor.java
@@ -23,9 +23,9 @@ import org.omg.CORBA.ValueMember;
 
 import java.io.IOException;
 import java.util.BitSet;
-import java.util.Map;
 import java.util.logging.Logger;
 
+import static org.apache.yoko.rmi.impl.FieldDescriptor.ValueMemberAccess;
 import static org.omg.CORBA.TCKind._tk_abstract_interface;
 import static org.omg.CORBA.TCKind._tk_alias;
 import static org.omg.CORBA.TCKind._tk_any;
@@ -83,33 +83,7 @@ class ValueMemberFieldDescriptor extends FieldDescriptor {
     private final ValueMember valueMember;
     private final TypeCode typeCode;
     private final TCKind kind;
-
-    /**
-     * Represents the access level of a ValueMember field.
-     */
-    enum ValueMemberAccess {
-        PRIVATE(0),
-        PUBLIC(1);
-
-        private final short value;
-
-        ValueMemberAccess(int value) {
-            this.value = (short) value;
-        }
-
-        static ValueMemberAccess fromShort(short access) {
-            for (ValueMemberAccess vma : values()) {
-                if (vma.value == access) {
-                    return vma;
-                }
-            }
-            throw new IllegalArgumentException("Invalid ValueMember access value: " + access);
-        }
-
-        boolean isPublic() {
-            return this == PUBLIC;
-        }
-    }
+    private final ValueMemberAccess access;
 
     /**
      * Creates a ValueMemberFieldDescriptor from a ValueMember.
@@ -123,7 +97,7 @@ class ValueMemberFieldDescriptor extends FieldDescriptor {
         this.valueMember = valueMember;
         this.typeCode = valueMember.type;
         this.kind = typeCode.kind();
-        this.isPublic = ValueMemberAccess.fromShort(valueMember.access).isPublic();
+        this.access = ValueMemberAccess.fromShort(valueMember.access);
     }
 
     @Override
